+++ src/dpr/components/async-request-list/clientClass.mjs
@@ -7,7 +7,7 @@ export default class DprAsyncRequestList extends DprPollingStatusClass {
   }
 
   initialise () {
-    this.POLLING_STATUSES = ['FINISHED', 'FAILED', 'EXPIRED']
+    this.END_STATUSES = this.getEndStatuses()
     this.POLLING_FREQUENCY = this.getPollingFrquency()
 
     this.requestList = document.getElementById('dpr-async-request-component')
@@ -22,9 +22,14 @@ export default class DprAsyncRequestList extends DprPollingStatusClass {
         const meta = JSON.parse(this.requestData)
         await Promise.all(
           meta.map(async (metaData) => {
-            const response = await this.getRequestStatus(metaData, this.csrfToken)
-            if (this.POLLING_STATUSES.includes(response.status)) {
-              window.location.reload()
+            // Don't poll if current state is an end state
+            if (!this.END_STATUSES.includes(metaData.status)) {
+              const response = await this.getRequestStatus(metaData, this.csrfToken)
+
+              // Reload if new status is an end state
+              if (this.END_STATUSES.includes(response.status)) {
+                window.location.reload()
+              }
             }
           }),
         )
