+++ src/dpr/utils/renderAsyncReport.ts
@@ -31,19 +31,23 @@ export const initDataSources = ({ req, res, services }: AsyncReportUtilsParams) 
     },
   )
   const summaryDataPromise = reportDefinitionPromise.then(
-    (definition: components['schemas']['SingleVariantReportDefinition']) =>
-      Promise.all(
-        definition.variant.summaries.map((summary) => {
-          return services.reportingService
-            .getAsyncSummaryReport(token, reportId, reportVariantId, tableId, summary.id, {
-              dataProductDefinitionsPath,
-            })
-            .then((data: Array<Dict<string>>) => ({
-              ...summary,
-              data,
-            }))
-        }),
-      ),
+    // eslint-disable-next-line consistent-return
+    (definition: components['schemas']['SingleVariantReportDefinition']) => {
+      if (definition.variant.summaries) {
+        return Promise.all(
+          definition.variant.summaries.map((summary) => {
+            return services.reportingService
+              .getAsyncSummaryReport(token, reportId, reportVariantId, tableId, summary.id, {
+                dataProductDefinitionsPath,
+              })
+              .then((data: Array<Dict<string>>) => ({
+                ...summary,
+                data,
+              }))
+          }),
+        )
+      }
+    },
   )
   const reportDataCountPromise = services.reportingService.getAsyncCount(token, tableId)
   const stateDataPromise = services.asyncReportsStore.getReportByTableId(tableId)
